+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The main work (all changes without a GitHub username in brackets in the below li
     - Enhancement: Re-establish subscriptions in parallel per peer on device/bridge startup
     - Fix: Ensures to report all attribute changes later that happened during an initial subscription seeding when dataVersion filtering was used
     - Fix: Only exports atomic-commands in Thermostat cluster server when relevant features are supported
+    - Fix: Properly cancels subscriptions that were canceled by the peer but were still in resubmission state
 
 - @matter/protocol
     - Breaking: Removed automatic retry-logic for interactions on node-reachabiloity issues, new session will be initialized automatically afterwards